+++ pkg/distributor/distributor_test.go
@@ -1444,6 +1444,91 @@ func TestDistributor_Push_ExemplarValidation(t *testing.T) {
 	}
 }
 
+func TestDistributor_ExemplarValidation(t *testing.T) {
+	tests := map[string]struct {
+		minExemplarTS     int64
+		req               *mimirpb.WriteRequest
+		expectedExemplars []mimirpb.PreallocTimeseries
+	}{
+		"valid exemplars": {
+			minExemplarTS: 0,
+			req: &mimirpb.WriteRequest{Timeseries: []mimirpb.PreallocTimeseries{
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test1"}, 1000, []string{"foo", "bar"}),
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test2"}, 1000, []string{"foo", "bar"}),
+			}},
+			expectedExemplars: []mimirpb.PreallocTimeseries{
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test1"}, 1000, []string{"foo", "bar"}),
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test2"}, 1000, []string{"foo", "bar"}),
+			},
+		},
+		"one old, one new, separate series": {
+			minExemplarTS: 300000,
+			req: &mimirpb.WriteRequest{Timeseries: []mimirpb.PreallocTimeseries{
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test"}, 1000, []string{"foo", "bar"}),
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test"}, 601000, []string{"foo", "bar"}),
+			}},
+			expectedExemplars: []mimirpb.PreallocTimeseries{
+				{TimeSeries: &mimirpb.TimeSeries{
+					Labels:    []mimirpb.LabelAdapter{{Name: model.MetricNameLabel, Value: "test"}},
+					Exemplars: []mimirpb.Exemplar{},
+				}},
+				makeExemplarTimeseries([]string{model.MetricNameLabel, "test"}, 601000, []string{"foo", "bar"}),
+			},
+		},
+		"multi exemplars": {
+			minExemplarTS: 300000,
+			req: &mimirpb.WriteRequest{Timeseries: []mimirpb.PreallocTimeseries{{
+				TimeSeries: &mimirpb.TimeSeries{
+					Labels: []mimirpb.LabelAdapter{{Name: model.MetricNameLabel, Value: "test"}},
+					Exemplars: []mimirpb.Exemplar{
+						{Labels: []mimirpb.LabelAdapter{{Name: "foo", Value: "bar1"}}, TimestampMs: 1000},
+						{Labels: []mimirpb.LabelAdapter{{Name: "foo", Value: "bar2"}}, TimestampMs: 601000},
+					},
+				}},
+			}},
+			expectedExemplars: []mimirpb.PreallocTimeseries{{
+				TimeSeries: &mimirpb.TimeSeries{
+					Labels: []mimirpb.LabelAdapter{{Name: model.MetricNameLabel, Value: "test"}},
+					Exemplars: []mimirpb.Exemplar{
+						{Labels: []mimirpb.LabelAdapter{{Name: "foo", Value: "bar2"}}, TimestampMs: 601000},
+					},
+				}},
+			},
+		},
+		"one old, one new, same series": {
+			minExemplarTS: 300000,
+			req: &mimirpb.WriteRequest{Timeseries: []mimirpb.PreallocTimeseries{{
+				TimeSeries: &mimirpb.TimeSeries{
+					Labels: []mimirpb.LabelAdapter{{Name: model.MetricNameLabel, Value: "test"}},
+					Exemplars: []mimirpb.Exemplar{
+						{Labels: []mimirpb.LabelAdapter{{Name: "foo", Value: "bar1"}}, TimestampMs: 1000},
+						{Labels: []mimirpb.LabelAdapter{{Name: "foo", Value: "bar2"}}, TimestampMs: 601000},
+					},
+				}},
+			}},
+			expectedExemplars: []mimirpb.PreallocTimeseries{{
+				TimeSeries: &mimirpb.TimeSeries{
+					Labels: []mimirpb.LabelAdapter{{Name: model.MetricNameLabel, Value: "test"}},
+					Exemplars: []mimirpb.Exemplar{
+						{Labels: []mimirpb.LabelAdapter{{Name: "foo", Value: "bar2"}}, TimestampMs: 601000},
+					},
+				}},
+			},
+		},
+	}
+	ds, _, _ := prepare(t, prepConfig{
+		numDistributors: 1,
+	})
+	now := mtime.Now()
+	for testName, tc := range tests {
+		t.Run(testName, func(t *testing.T) {
+			err := ds[0].validateSeries(now, tc.req.Timeseries[0], "user", false, tc.minExemplarTS)
+			assert.NoError(t, err)
+			assert.Equal(t, tc.expectedExemplars, tc.req.Timeseries)
+		})
+	}
+}
+
 func BenchmarkDistributor_Push(b *testing.B) {
 	const (
 		numSeriesPerRequest = 1000
@@ -2679,16 +2764,19 @@ func makeWriteRequestHA(samples int, replica, cluster string) *mimirpb.WriteRequ
 func makeWriteRequestExemplar(seriesLabels []string, timestamp int64, exemplarLabels []string) *mimirpb.WriteRequest {
 	return &mimirpb.WriteRequest{
 		Timeseries: []mimirpb.PreallocTimeseries{
-			{
-				TimeSeries: &mimirpb.TimeSeries{
-					// Labels: []mimirpb.LabelAdapter{{Name: model.MetricNameLabel, Value: "test"}},
-					Labels: mimirpb.FromLabelsToLabelAdapters(labels.FromStrings(seriesLabels...)),
-					Exemplars: []mimirpb.Exemplar{
-						{
-							Labels:      mimirpb.FromLabelsToLabelAdapters(labels.FromStrings(exemplarLabels...)),
-							TimestampMs: timestamp,
-						},
-					},
+			makeExemplarTimeseries(seriesLabels, timestamp, exemplarLabels),
+		},
+	}
+}
+
+func makeExemplarTimeseries(seriesLabels []string, timestamp int64, exemplarLabels []string) mimirpb.PreallocTimeseries {
+	return mimirpb.PreallocTimeseries{
+		TimeSeries: &mimirpb.TimeSeries{
+			Labels: mimirpb.FromLabelsToLabelAdapters(labels.FromStrings(seriesLabels...)),
+			Exemplars: []mimirpb.Exemplar{
+				{
+					Labels:      mimirpb.FromLabelsToLabelAdapters(labels.FromStrings(exemplarLabels...)),
+					TimestampMs: timestamp,
 				},
 			},
 		},